chore(jazz-run): split commands and core logic and add tests
